polymarket-whale-detector/requirements.txt
# Python 3.11+
httpx[http2]>=0.27.0
python-telegram-bot>=21.0
supabase>=2.5.0
python-dotenv>=1.0.0
py-clob-client>=0.34.0
